+++ src-ui/js/ui/Listener.js
@@ -52,7 +52,10 @@ ui.listener = {
 		}
 
 		ui.menuconfig.sync();
-		ui.menuconfig.set("autocheck_once", ui.menuconfig.get("autocheck"));
+		ui.menuconfig.set(
+			"autocheck_once",
+			ui.menuconfig.get("autocheck_mode") !== "off" && ui.puzzle.playeronly
+		);
 		ui.currentpid = puzzle.pid;
 
 		ui.adjustcellsize();