+++ src/variety/roma.js
@@ -338,7 +338,7 @@
 	FailCode: {
 		bkDupNum: [
 			"1つの領域に2つ以上の同じ矢印が入っています。",
-			"An area has plural same arrows."
+			"An area has duplicate arrows."
 		],
 		stopHalfway: [
 			"ゴールにたどり着かないセルがあります。",