+++ src/variety/shugaku.js
@@ -354,7 +354,7 @@
 		paint: function() {
 			this.drawBGCells();
 			this.drawShadedCells();
-			this.drawDotCells(false);
+			this.drawDotCells();
 			this.drawDashedGrid();
 
 			this.drawFutons();