+++ test/script/doppelblock.js
@@ -1,0 +1,30 @@
+ui.debug.addDebugData("doppelblock", {
+	url: "5/5/g6g34g10h",
+	failcheck: [
+		[
+			"ceNoNum",
+			"pzprv3/doppelblock/5/5/. . 6 . 3 4 /. . . . . . /1 . . . . . /0 . . . . . /. . . . . . /. . . . . . /"
+		],
+		[
+			"nmDupRow",
+			"pzprv3/doppelblock/5/5/. . 6 . 3 4 /. . . . . . /1 - 1 - 1 2 /0 . . . . . /. . . . . . /. . . . . . /"
+		],
+		[
+			"ceTooManyBlocks",
+			"pzprv3/doppelblock/5/5/. . 6 . 3 4 /. . . . . . /1 . . . . . /0 - - 1 - 2 /. . . . . . /. . . . . . /"
+		],
+		[
+			"nmSum",
+			"pzprv3/doppelblock/5/5/. . 6 . 3 4 /. . . . . . /1 - 1 2 - 3 /0 . . . . . /. . . . . . /. . . . . . /"
+		],
+		[
+			"nmSum",
+			"pzprv3/doppelblock/5/5/. . 6 . 3 4 /. . . . . . /1 - 1 - 2 3 /0 1 - 3 - 2 /. . . . . . /. . . . . . /"
+		],
+		[
+			null,
+			"pzprv3/doppelblock/5/5/. . 6 . 3 4 /. 3 - 2 1 - /1 - 1 - 2 3 /0 2 3 - - 1 /. - 2 1 3 - /. 1 - 3 - 2 /"
+		]
+	],
+	inputs: []
+});